Dedupe bulk-upload dialogs into one generic + thin wrappers
BulkUploadLlmItemsDialog and BulkUploadLlmGeneralItemsDialog were ~90%
identical (CSV parse, variable columns, annotation columns, upload, preview
grid, guidelines). Extract all of that into a single generic
BulkUploadItemsDialog driven by a small content-column adapter.

- New BulkUploadItemsDialog owns the shared machinery.
- Each dialog is now a thin wrapper that declares only its content columns
  (conversation_history + agent_response vs input + output), sample rows, and
  copy. Public props are unchanged, so the task page is untouched.

Behaviour preserved; minor harmless tweaks: required-field error messages are
now per-column, and llm-general items now always send evaluator_variables
(matching the LLM dialog).
